+++ pom.xml
@@ -25,6 +25,9 @@
     <allure-maven.version>2.12.0</allure-maven.version>
     <appium.java.version>8.3.0</appium.java.version>
     <lombok.version>1.18.20</lombok.version>
+    <surefire.version>3.0.0-M5</surefire.version>
+    <xml.path>src/test/resources/testng-all.xml</xml.path>
+    <mvn.compiler.version>3.8.1</mvn.compiler.version>
     <maven.compiler.source>11</maven.compiler.source>
     <maven.compiler.target>11</maven.compiler.target>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -118,15 +121,15 @@
     <plugin>
       <groupId>org.apache.maven.plugins</groupId>
       <artifactId>maven-surefire-plugin</artifactId>
-      <version>3.0.0-M5</version>
+      <version>${surefire.version}</version>
     </plugin>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.8.1</version>
+        <version>${mvn.compiler.version}</version>
         <configuration>
-          <source>9</source>
-          <target>9</target>
+          <source>10</source>
+          <target>10</target>
         </configuration>
       </plugin>
     </plugins>
