+++ src/main/java/com/epam/training/student_mykola_koltutskyi/enums/MobilePlatformType.java
@@ -1,0 +1,8 @@
+package com.epam.training.student_mykola_koltutskyi.enums;
+
+
+
+public enum MobilePlatformType {
+    ANDROID,
+    IOS
+}
