+++ src/main/java/com/epam/training/student_mykola_koltutskyi/models/Form.java
@@ -1,0 +1,15 @@
+package com.epam.training.student_mykola_koltutskyi.models;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class Form {
+    private final String numberOfInstances;
+    private final String machineType;
+    private final String gpuModel;
+    private final String localSSD;
+    private final String dataCenterLocation;
+    private final String estimatedCost;
+}
