+++ src/main/java/com/epam/training/student_mykola_koltutskyi/pages/AbstractPage.java
@@ -1,0 +1,11 @@
+package com.epam.training.student_mykola_koltutskyi.pages;
+
+import org.openqa.selenium.WebDriver;
+
+public abstract class AbstractPage {
+    protected WebDriver driver;
+
+    protected AbstractPage(WebDriver driver) {
+        this.driver = driver;
+    }
+}
