+++ src/main/java/com/epam/training/student_mykola_koltutskyi/pages/EmailGeneratorPage.java
@@ -1,0 +1,22 @@
+package com.epam.training.student_mykola_koltutskyi.pages;
+
+import lombok.Getter;
+import org.openqa.selenium.WebDriver;
+
+public class EmailGeneratorPage extends AbstractPage {
+    private final WebDriver driver;
+    @Getter
+    String generatedEmail;
+
+    public EmailGeneratorPage(WebDriver driver) {
+        super(driver);
+        this.driver = driver;
+    }
+
+    public void generateEmail() {}
+
+    public CalculatorPage returnToCalculator() {
+        return new CalculatorPage(driver);
+    }
+
+}
