+++ src/main/java/com/epam/training/student_mykola_koltutskyi/service/DriverData.java
@@ -1,0 +1,13 @@
+package com.epam.training.student_mykola_koltutskyi.service;
+
+import lombok.Builder;
+import lombok.Getter;
+import com.epam.training.student_mykola_koltutskyi.enums.BrowserType;
+import com.epam.training.student_mykola_koltutskyi.enums.MobilePlatformType;
+
+@Builder
+@Getter
+public class DriverData {
+    private BrowserType browserType;
+    private MobilePlatformType platformType;
+}
