--- src/main/java/org/yandrut/enums/BrowserType.java
@@ -1,8 +1,0 @@
-package org.yandrut.enums;
-
-public enum BrowserType {
-    CHROME,
-    FIREFOX,
-    EDGE
-
-}
