--- src/main/java/org/yandrut/pages/AbstractPage.java
@@ -1,11 +1,0 @@
-package org.yandrut.pages;
-
-import org.openqa.selenium.WebDriver;
-
-public abstract class AbstractPage {
-    protected WebDriver driver;
-
-    protected AbstractPage(WebDriver driver) {
-        this.driver = driver;
-    }
-}
