--- src/main/java/org/yandrut/selenium/manager/FirefoxManager.java
@@ -1,13 +1,0 @@
-package org.yandrut.selenium.manager;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
-
-public final class FirefoxManager {
-
-    private FirefoxManager() {}
-
-    public static WebDriver getDriver() {
-        return new FirefoxDriver();
-    }
-}
