+++ src/main/resources/main.properties
@@ -1,10 +1,13 @@
 data.form.NUMBER_OF_INSTANCES = 4
-data.form.MACHINE_TYPE = n1-standart-8
+data.form.MACHINE_TYPE = n1-standard-8
 data.form.GPU_TYPE = NVIDIA V100
 data.form.LOCAL_SSD = 2x375 GB
 data.form.DATA_CENTER_LOCATION = Netherlands (europe-west4)
-data.form.ESTIMATED_COST = $1,820.80
+data.form.ESTIMATED_COST = $8,040.40
 data.provided.cloudUrl = https://cloud.google.com
 data.provided.calculatorUrl = https://cloud.google.com/products/calculator
 data.provided.searchPrompt = Google Cloud Pricing Calculator
-data.provided.expectedPageName=Cloud Computing Services | Google Cloud
+data.provided.expectedPageName = Cloud Computing Services | Google Cloud
+data.provided.downloadPath = /target/download
+data.provided.tabIndex = 1
+data.provided.tabInfo = Google Cloud Estimate Summary
