--- src/main/resources/testng-all.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-
-<suite name="All tests">
-    <test name="All">
-        <classes>
-            <class name="org.yandrut.SmokeTests">
-                <methods>
-                    <include name="redirectsToCalculatorPage"/>
-                    <include name="priceIsGettingEstimated"/>
-                </methods>
-            </class>
-        </classes>
-    </test>
-</suite>
