+++ src/test/resources/testng-all.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+
+<suite name="All tests">
+    <test name="All">
+        <classes>
+            <class name="com.epam.training.student_mykola_koltutskyi.AllTest">
+                <methods>
+                    <include name="redirectsToCalculatorPage"/>
+                    <include name="priceIsGettingEstimated"/>
+                    <include name="redirectsToDetailedView"/>
+                    <include name="correctDataProvided"/>
+                </methods>
+            </class>
+        </classes>
+    </test>
+</suite>
